Security_Scripts/qrng_steganography/requirements.txt
Pillow==10.3.0
numpy==1.26.4
cryptography==42.0.5
